requirements.txt
numpy==1.16.2
requests==2.26.0
pyhocon==0.3.58
torchvision==0.8.2+cu110
torch==1.7.1+cu110
Pillow==8.4.0
termcolor==1.1.0